Revert "SeedMechanisms/Glarborg/C2 now at v3"
This reverts commit caa49b14bcb45f2efec19566d7c13b4ec344ffa9.

As Franklin noted in the comment at
http://github.com/rwest/RMG-Java/commit/caa49b14
the changes in that commit are valid only for fuel-lean low-temperature
oxidation (i.e. below 800 K). The 8 species removed are irrelevant for low
temperature oxidation, and including them in the mechanism only bogs things
down, so if this matches your conditions you may want the change, but in
the interests of generality, I'm reverting the commit.
(to get it back do 'git cherry-pick caa49b14b')