Adding forbidden groups for reverse reaction templates
RMG allows a user to define forbiddenGrooups for the forward reaction
templates (those found in the kinetics_groups directory), while the
forbiddenGroups for any reverse reaction template was null.

Now, the forbiddenGroups for the reverse reaction template are assumed to
be the forbiddenGroups of the forward reaction template.

Using this new feature, MRH has added O=O as a forbidden group to the new
Disproportionation family, so that O=O + R-H --> HO2 + R* cannot occur
via this reaction family (the structure of this reaction is generated by
the H_abstraction reaction template, with O=O replaced by *O-O*).  With
JDM having already added HO2 as a forbidden group (to form O=O), the new
oxygen database should be complete / resolved.

Closes #55