HO2+OH=H2O+O2 rates changed in Glarborg reaction library.
Glarborg expressed this rate as the sum of three modified Arrhenius
expressions, one of which has a negative A factor. RMG does not like
this negative A factor so I have replaced the expression with one
from Baulch et al. Perhaps there is a more appropriate rate (one with
a larger valid range of T), or one could re-fit the Glarborg to a
single Arrhenius, but this will at least allow RMG to run...