Allow :CO insertion into CsO--H in 1,2_insertion family.
This allows methlyformate to decompose into CO and CH3OH.
The pathway was proposed by Dr Nabiha Chaumeix and the rate
was calculated by C.F. Goldsmith in January 2010 while
R.H.West was visiting Nabiha in Orleans (France). But for
some reason it was not added to the version of the database
under version control.

I think by now Franklin has probably done a higher level
calculation, but this is the one we had back then.