Updated Methylformate Reaction Library with Franklin's latest rates
This message from CFG:

I redid the rates for the decomposition direction. I found a symmetry number
error in one of the files, so in paranoia I redid all of them. Lastly, I did
four of the H abstraction rates. On the good side, it is four of the six most
important. On the bad side, #2 & #3 are missing. For some reason, I cannot get
the OH and O transition states to converge. I have H and CH3; they are done
using CBS-QB3, since the QCI method takes too long. Let me know how it makes a
difference!