BUG FIX: Removing Y_2centeradjbirad from Y_rad_birad Union
The Y_2centeradjbirad node is no longer a direct child of Y_rad_birad.
Instead, it is a child of Y_rad.  Although this new definition was fixed in
the tree.txt file, it was not fixed in the dictionary.txt file.

The problem: RMG would incorrectly estimate all .O-O. + R-H --> R. + HO-O.
pre-exponential factors by a factor of 2.  For instance, imagine O2 +
isobutene.  There are 6 primary hydrogens and two radical sites on O2, meaning
the A stored in the RMG database should be multiplied by 12.  However, Yushi
Suzuki pointed out that RMG was returning a chem.inp file in which the A was
multiplied by 24.  The reason for this was RMG finding 6x2 combinations when O2
acted as "Y_rad" and another 6x2 combinations when O2 acted as
"Y_2centeradjbirad".

This commit resolves the problem.