Added rate estimate for X_H + .OO. = X. + HOO. abstraction.
I used Bill's estimate that Mike had earlier assigned to
ROOH + .OO. = ROO. + HOO. abstractions, and generalized it.
Anything that *has* a defined rate will not be changed by
putting such an estimate at the top of the tree, but things
without it (like .ROH + .OO. = .RO. + HOO.) will now use
this instead of a crazy average with a fixed, low, barrier that
is often less than the endothermicity.

Also moved the restructuredtext macros to bottom of the comments file.