Read duplicate PLOG rates from Seed Mechanisms etc.
When reading PLOG reactions from a Seed Mechanism (or Reaction Library etc. as these
re-use this SeedMechanism methods) we check if we have already read in the same
reaction *from this seed mechanism* and if so we append the rates as duplicates.
This is useful for mechanisms such as Glarborg's where things can be expressed as
the sum of multiple PLOG reactions.

It currently cannot cope well if two such sets have the same high pressure limits,
as the Reaction.addAdditionalKinetics method just edits the redundancy in this case,
and writing the chemkin string at the end can't find the corresponding high P rate.
As a work-around I have edited the Glarborg mechanisms to avoid this
(with no change to the chemistry).