A few bugfixes to the H abstraction dictionary.
The most important is fixing one adjacency list so that it is valid.
Previously it referred to an atom that did not exist and not to an atom
that did exist. Not sure why RMG-Java didn't crash on this one before.

Also fixed a number of bond close brackets (were ')', should be '}').
This probably didn't affect RMG Java at all, but is still a good thing
to do for the sake of consistency.