Removed non-numeric indices from H abstraction rate library.
RMG Java seems to be able to handle non-numeric values for the index.
However, these are the only two places where that is taken advantage of.
Instead, I'd rather force all indices to be numbers (preferably
integers). For lack of a better choice, the two affected entries became
1001 and 1002.

Also updated the relevant comments to use the new indices.