Some fixes to the Dooley methylformate kinetics libraries.
The reaction H + CO2 = OCHO was listed in both the pressure-independent
and pressure-dependent sections with the same kinetics, despite their
having vastly different meanings (and units!) in both places. After
consulting the reference we made a best guess as to which is correct
(the pressure-dependent one) and deleted the other.

Also changed the A-factor of CH2O + HCO = CH2OCHO to match what was in
the same reference; it was off by an order of magnitude. The reference
suggests that they had to fit this value to their data, which implies
that this reaction is rather important (and that it is very important
for these significant reactions that you check their kinetic parameters
very closely!) Note that we got the original value from their Chemkin
file in the Supporting Materials. If this commit makes models worse we
should follow up with Dooley et al.