Fixed a few malformed adjacency lists in the Dooley kinetics libraries.
It's a bit disconcerting that RMG-Java will just carry on when this is
encountered. I'm not sure if it just dropped the species, or tried to
guess which is correct, or exactly what the outcome was. RMG-Py will
fail loudly in this situation, which is how I found the bug in the
first place.