Removed a few unexpected duplicate reactions from Glarborg C3 library.
The reactions H2CCCH2 + H = CH3CHCH, H2CCCH2 + H = CH3CCH2, and
H2CCCH2 + H = CH2CHCH2 have entries in both the pressure-independent and
pressure-dependent sections of the Glarborg C3 mechanism. (This was only
discovered when trying to import to the new database format, for which
my script checks for duplicates between these sections, while Chemkin
does not.) Checking the reference, it seems the values given in the
pressure-independent section are for 0.1 bar only, and the reference
also has Arrhenius parameters at several other pressures.

Since we already have Troe expressions for these reactions in the
pressure-dependent sections, and since the low-P Arrhenius expression
in the Troe form matches those from the pressure-independent section,
we decided to keep the Troe forms and comment out the 0.1 bar Arrhenius
parameters.