A few bugfixes found in diff comparsion to official RMG-Java database.
Using recursive diff to compare the database in the RMG-Java repository
to this database revealed a number of incorrectly commented out
DUPLICATE tags, plus an incorrect structure. These are probably the
result of improper handling of a cherry-picked commit when a conflict
arose.

These fixes ensure that, after the extensive cherry-picking of commits,
we have arrived at the same database we had in the RMG-Java repository
as of today. We have also managed to preserve the history of the
database only.