Renamed databases folder to output.
We are about to convert the database into a new Python-based format that
is much more flexible and robust than the current Java-style syntax.
Eventually that format will be the only one kept under version control,
but for now we expect to keep both formats under version control for a
while. The idea is to add data to the new database format (since it is
better in many ways), and then export to the old format for use in
RMG-Java. Renaming the databases folder to output better reflects this
intended workflow.