Added first version of script used to import to the new database format.
An important step in switching database formats is to migrate all of the
existing data to the new format. This script is used to do exactly that.
Most of the magic happens in the database classes in RMG-Py; this script
just calls them in the appropriate order. This script also tags each
entry in the imported database with a note in its history indicating the
time of the import (and that it was by me! bwahaha).