Imported old data into new database format.
The new database format is based on Python syntax -- in fact, each file
in the new database is a valid Python file, and is therefore named with
the .py extension. All database files are encoded in UTF-8 to allow for
non-ASCII characters; be sure you use a text editor that respects this
(i.e. not regular Notepad, Windows users). The encoding is indicated
under the shebang at the top of each file so that the exec() command in
Python respects the encoding.

All of the thermodynamics, kinetics, and frequencies groups and
libraries were imported. The transport database and liquid-phase
database were not imported, as these features are not yet implemented in
RMG-Py.

Users should now be adding entries to the new Python-style database,
because it is more flexible and robust. An export script will soon be
added that enables exporting of the Java-style database for use in
RMG-Java.

This is simply the result of running the importOldDatabase.py script.
There is still some manual cleaning up to do, particularly in adding
reference information, but the most important thing is that the data is
preserved.