Added script for exporting to the old database format.
The new script exportOldDatabase.py does the opposite of its counterpart
importOldDatabase.py: it takes the new database format and exports it
into the old format so that we can use it in RMG-Java. Again, most of
the magic happens in the database classes in RMG-Py. Note that, as the
export script produces the old database automatically, it will likely be
organized a bit differently than the hand-curated one; it will certainly
have lost all comments (and commented-out entries).