First export of old database format from new database.
This was done right away so that, when it is repeated in the future, it
is easier to diff the resulting output to see what changed. Note that
all of the inline comments that did not get imported to the new database
have been deleted due to the automatic nature of the export. All of
these comments are still available in the repository history, of course,
if you ever want to see what they were. (That was one of the reasons we
wanted to preserve the revision history of the old database.)

On the flip side, the database files should now look a lot neater due to
the use of nice string formatting. All use of tabulators should also be
replaced with spaces.

Also note the change of units for the kinetics databases, which now use
mol/m3/s instead of mol/cm3/s (since that's what RMG-Py uses
internally).