Added new version of GRI-Mech 3.0 thermo library with NASA polynomials.
The version of GRI-Mech 3.0 imported from the old RMG-Java database used
the ThermoData thermo model with 7 heat capacities. In reality, GRI-Mech
provides NASA polynomials for this data. Since we can store NASA
polynomials in the new database format, it is silly to keep the values
from the old database. This commit provides a version of GRI-Mech 3.0
with NASA polynomials.

This is one demonstration the versatility of the new database format:
storing thermo data in multiple formats.