Added kinetics depositories imported from PrIMe data warehouse.
These depositories were automatically generated using a script that
crawls the PrIMe database, determines the reactants and products for
each reaction, and attempts to see if it can fit that reaction uniquely
to one (and only one) of RMG's reaction families. If so, all of the
kinetics for that reaction are placed into the depository for the
appropriate family as separate entries. Note that this only works for
C/H/O chemistry where the reactants and products are real species and
the kinetics is an Arrhenius expression (rather than a branching ratio).

This provides lots of training/test data for some reaction families,
gets us started in several others, and leaves out some as well. From the
current PrIMe database, 2652 entries from 1075 unique reactions were
identified as being part of only one reaction family. Of these,
nearly half (1252) came from the H_Abstraction family alone. The other
major contributors were R_Recombination (601) and
R_Addition_MultipleBond (407). Several families had no hits in the PrIMe
database, and no PrIMe depository components were added for those
families.