Removed duplicate entry with wrong kinetics from H_Abstraction depository.
When I looked up the deleted entry in the reference, its kinetics were
quite different than the ones given. However, the very next entry in the
depository was for the same reaction and reference, but with the
correct kinetics. Thus I deleted the first entry, as it was a duplicate
of the second but with wrong kinetics.