Fixed a few cases of wrong reaction direction in H_Abstraction depository.
In PrIMe the forward and reverse directions are stored together under
the same PrIMe ID. The kinetics then have a flag indicating whether they
describe the forward or reverse process. These entries presumably have
that flag set to the wrong value, as the references suggest that the
kinetics are defined in the opposite direction from the one indicated.
To fix this I have simply exchanged the reactants and products.