Some bugfixes to the H_Abstraction depository from MRH.
MRH did some validation of the PrIMe data for H_Abstraction a while
back. This commit incorporates some of his finds, some of which are
rather subtle. Many of these points were not showing up as outliers on
my group additivity parity plots, but it is nonetheless important to
have everything be correct, even the points that don't stand out as
different.