Removed "X_H_or_Xrad_H;Y_rad_birad" rate rule for H abstraction.
The X_H_or_Xrad_H node is defined as a group in the H abstraction groups,
but it is not in the tree, so this rule is probably being ignored. Since
it's not in the tree, my group additivity value fitting code is choking on
this entry. As the note is not being used anyway, I am removing it for now.
If we decide we want it back, we should actually amend the tree instead.