Added script for generation of kinetics group additivity values.
This script is currently designed more for evaluating the group additivity
values rather than saving them to the database. The hardest part is getting
the training and test set into the appropriate direction and with the
necessary information in order to be used to train and test the groups. In
particular, some of the training and test reactions are given in the
reverse of the direction for which the family is defined, which requires
generating the reverse kinetics (which itself requires thermodynamics).

Currently the script fits k(T) values at each temperature; in the future we
will likely explore alternative fitting algorithms.