Renamed generateKineticsGroupValues.py to kineticsGroups.py.
This is anticipation of this script being used for multiple purposes.
Currently the script is designed more for testing and evaluating the
group additivty values by producing parity plots. We also want to be
able to generate and save the group values for all the families, so that
we can use them in RMG-Py. The latter functionality is coming soon.