Fixed some groups in kinetics rate rules for intra_H_migration.
There are several nodes in the old intra_H_migration tree below
Cs_H_out_H/NonDeC that are commented out. The comments suggest that these
particular nodes do not conform to the mutual exclusivity requirements of
our tree structure, and that's why they are commented out. (We tried to
fix them in April 2010, but gave up.) However, there are a bunch of rate
rules that still use these groups. It's not clear what RMG-Java was doing
with these, but their existence was causing the group additivity
generation scheme to fail for intra_H_migration.

To fix them, I changed all of the rate rules to use the parent node
Cs_H_out_H/NonDeC instead, at least for the short term.