Added shebang and encoding comments to PrIMe sections of kinetics depository.
These are necessary for the codecs module to be able to properly load and
save the database with non-ASCII characters. Unfortunately I forgot to add
them as part of the script that exported the PrIMe database entries into
our format. The effect of omitting these is a UnicodeDecodeError or
UnicodeEncodeError raised when processing a unicode string containing a
non-ASCII character.