Marked all shortDesc and longDesc strings in database as unicode strings.
I had erroneously hoped that Python would automatically treat all strings
as unicode strings in source files marked with the UTF-8 encoding. Clearly
that is incorrect. There were a few longDesc strings in the
R_Recombination/rules kinetics depository that had unicode characters, and
this was preventing the loaded database from being resaved due to a
UnicodeDecodeError, since Python thought it was simply an ASCII string.

Fortunately, all I needed to do was to mark the offending strings as
unicode strings and reload and resave with the RMG-Py database output code
updated to mark all shortDesc and longDesc strings as unicode (i.e. with
the 'u' prefix). The diff of this commit actually reflects this for the
most part. (A few adjacency lists were reordered, along with a few
parameter lists, but again this is clear from the diff output.)