Added functionality for generating and saving kinetics group values to script.
There are at least two use cases for the "generate kinetics group additivity
values" functionality: (1) generate and save the group values for all
reaction families for use in RMG jobs and (2) generate and evalutate the
group values for one or more individual families. Since a lot of the code
that performs the kinetics group additivity value generation is the same in
either case, I've placed functionality for both in the same file. You can
invoke either using the command line, i.e.

$ python kineticsGroups.py generate

to generate and save group values for all families, or

$ python kineticsGroups.py evaluate <family>

to generate and evaluate group values for a single family (as specified as
an argument). The argparse module makes this quite straightforward to set
up.