Fixed error in preexponential factors for R_Addition_MultipleBond reactions.
These three fixes all come from the same paper. In that paper, the Arrhenius
data is given using log A. As best I can tell, whoever put this into PrIMe
interpreted this as the natural logarithm (base e) instead of the common
logarithm (base 10). The result: an error of thirty orders of magnitude!

The activation energies were correct, but I converted them to units of
kcal/mol anyway because that's what is used in the reference.