Changed to high-P limit Arrhenius parameters for some R_Addition_MultipleBond reactions.
The reference for these reactions gives their kinetics at a series of
pressures and for two temperature ranges. Originally the parameters were
from the P = 1 bar values. We want to use high-P values instead for this
depository, so I've swapped those values in. I've also indicated the
temperature ranges so that each kinetics is only used to train group
values within its applicable range.

Also note that the activation energies of these entries was way too large
anyway, which is how I flagged them for checking in the first place.