Fixed even more incorrect Ea values in R_Addition_MultipleBond.
These 15 errors again come from within the same reference, and within the
same table in that reference (Table 3 on p. 436). I am starting to think
there is a bug in the way the Ea values were exported from PrIMe by my
script, since so many entries have been way off, and not in an obvious
way.