Do not train groups on Evans-Polanyi reaction rules when deltaH is unknown.
Because our reaction rules in the old database do not have the specific 
molecules, we do not know the enthalpy change, and cannot multiply out the
EP expression to find the correct kinetics. It is therefore unsafe to train
against them. (They'll mostly be higher level nodes and estamites anyway,
I'd imagine)