Added training and test set files to each family in kinetics depository.
These files provide a place for people to add new kinetics data to be used
for training kinetics group additivity values. Any new kinetics data should
be added using the real reactant and product molecules, as we can determine
the appropriate group values programmatically. This enables us to
reorganize and manipulate the tree independently, and still be able to use
the same data for training (and testing) group additivity values.

In general, if you have high-pressure limit kinetics data for a reaction
involving not-too-small molecules, then you should put your data in the
training set. If your reaction involves molecules that probably are not
characteristic of the reaction family (e.g. very small molecules) or if
your data is pressure-dependent, then it probably should go in the test
set. I expect that most people will be adding data to the training set.

Still to do: (1) Determine which rate rules are for real reactants and
products, and move them from rules to training/test accordingly; (2) also
use the training/test data when exporting the old RMG-Java database by
programmatically converting their entries to rate rules.