Added evansPolanyi.py script for generating Evans-Polanyi plots.
This script accepts the name of a kinetics depository (containing real
reactions, not rate rules), and generates an Evans-Polanyi plot from the
entries in that depository. For now, the reactions are kept in the
direction specified in the file rather than all being converted to the
forward direction for this family.

These plots are often useful for double-checking that the kinetic
parameters in a depository do not have any obvious errors.