Moved lots of H_Abstraction butanol reactions from rules to training set.
MRH computed kinetics for these reactions, but had to improvise a bit in
order to get them into the old RMG database format. In particular, he
placed "groups" in the H_Abstraction tree for all of the real species
needed by these reactions.

We now have a better way of handling this information: the training set
in the kinetics depository. By converting to use real species, we can
decouple the tree from the training data and thereby make it easy to
regenerate kinetics group additivity values when either the tree or the
training set are modified. Furthermore, we eliminate the awkward
construction of groups in the tree that are actually real molecules.

Don't forget to convert the kinetic parameters from per-site basis to
total rate and include the degeneracy as an additional attribute. Also
don't forget to change ArrheniusEP to Arrhenius during the conversion.