Moved a few more entries from H_Abstraction rules to training depository.
These entries were generated by SSH and MHS, generally using the same
approach as MRH. This should take care of most of the rules that we added
from Green group calculations for H_Abstraction, although there are a few
"estimate" rules that we should consider replacing or removing. Many of
the remaining rules that aren't from the Green group also involve real
species, and those can and should be moved at some point, although this is
a much more involved task.

The next step is to make it so we can still use the entries in the training
set to generate rate rules to export to the old database.