When exporting old database, also generate rate rules from training sets.
Recently we added the training set to each reaction family depository as a
place for users to add kinetics data. With this commit, these can be
converted to rate rules so that the old database can use them. This makes
automatic the determining of the groups for each reaction and the conversion
to a per-site basis.

Now there shouldn't be much (except perhaps a few bugs) stopping people
from adding data to the new database instead of the old one.