Can now handle multiple training and test sets when generating kinetics groups.
Each set is plotted in a different color. The training sets are plotted as
circles, the test sets as squares. A set of check boxes allows for showing
and hiding of individual sets, making the plots more interactive.

The default is now to use rules.py and training.py as the training set and
PrIMe.py and test.py as the test set.