Added fitGroupValues() method (moved here from KineticsGroups class).
This method provides group additivity fitting by directly fitting k(T)
values at each of a set of temperatures. This is by no means the only
method of generating group additivity kinetic parameters, and we expect
to implement more approaches very soon.

This is a direct move of the old KineticsGroups.fitGroupValues() method.
Some refactoring is in order.