Preparing kineticsGroups.py to work with multiple group fitting algorithms.
Mostly this is gathering of the pieces that depend on the currently
implemented approach (which fits k(T) at discrete temperature points) into
one place, and tweaking some of the rest of the script accordingly.
Notably, fitGroupValues() now accepts a `method` parameter that indicates
the method to use to fit the group values. This method is now also
responsible for storing the fitted group values on the appropriate
entries in the group database and for printing the summary of the fitted
group values.