Significant rewrite of kineticsGroups script.
The motivation for this rewrite was to enable the script to be much more
flexible in terms of what it can do, and what options it has. To that end,
the command-line interface is noticeably different (and more complex).

The other big change is that the generate and evaluate operations are now
completely separate. Before the group additivity values would be generated
even during the evaluate operation, which is silly to be doing over and
over again. To enable this, the generate operation now also saves the
group values if they have changed, in either KineticsData or Arrhenius
format depending on how the group values are fitted.

The separation of generation and evaluation also means that we don't need
a test set during the generation step, nor do we need a training set during
the evaluation step. This makes things simpler.

The java operation is (hopefully) unchanged from before.