Added preliminary plotting of parity plots using RMG-Java as model.
The goal is to be able to evaluate the kinetics estimation accuracy of
both the rate rules approach of RMG-Java and the group additivity approach
of RMG-Py, to see which is "better". The implementation assumes the
presence of *_RMG_Java.py files in the kinetics libraries that contain
the RMG-Java estimated kinetics for the same reactions as in the
corresponding depository. (The java operation of kineticsGroups.py can
be used to generate such files.)