kineticsGroups java script has more filtering.
It now filters out reactions in the wrong direction,
and by editing the code (read the comments) you can 
replace rates with those from rmg-py (so that you can
plot parity plots). It also skips over rates that 
the java does not need to use an "Average of" for,
so that you can tell which are which (edit the code to
include *only* these rates instead of excluding them).

Various versions of this script were used to generate the 
parity plots for the poster presented at the 7th International
Conference on Chemical Kinetics.