Renamed "test" kinetics depositories to "depository" for each family.
The new depository.py files are to be used for placing kinetics data from
the literature of any (decent) quality. From this set we will extract
trusted kinetics parameters to use to train group additivity values and
for kinetics estimation. If you don't know where to put the kinetics you've
found, put them in the depository.py file of the corresponding reaction
family.