Imported RMG-Java kinetics libraries using new importOldDatabase.py script.
These have comments and MultiKinetics (instead of duplicates).
One strange thing is that the adjacency lists no longer seem
to sort the atoms by connectivity, meaning that they now seem
to vary, making the diff a little confusing.

Also, the modified 'history' items make the diff annoying.
For those two reasons, this for now is on a separate branch,
not master.