Added rate rules to intra_H_migration from Sandeep calculations and Atkinson literature.
Sandeep calculations are taken from DOI: 10.1021/jp9098792

The A factor in the paper differs from the ones added to the library as the
paper assumes a slightly different Arrhenius form compared to the default RMG
one. A few new groups were added to the dictionary and tree files to account
for the change in the TS ring size or the H being abstracted. The new library
entries have been checked by performing PopulateReactions jobs with the
original molecules in Sandeep's paper and checking if the right entry is
located.

RMG's previous estimates were based on Curran's estimates for ROO.
isomerizations and were atleast 4 orders of magnitude too slow at about
1000 K when compared with recommendations from NIST, a review by Orlando
et al.(DOI:10.1021/cr020527p) and recent articles that I have looked at
(Veerecken2010(DOI: 10.1039/c0cp00387e) and Somnitz2008(DOI: 10.1039/b711429j)).

The rates added correspond to 1,5 shift only as these are believed to be
faster than others due to ring strain considerations. I have not added
special rates for migration in RC(O)CH2C(O.)R' type of species which can
arise in low temperature oxidation due to cleavage of the O-O bond in
ketohydroperoxides. This is based on a preliminary examination of the
results by Veerecken2010 who report that the presence of the carbonyl
does not significantly alter the rate compared to the base alkoxy radical.