Fixed bug in Intra_R_Add_Endocyclic group definitions.
The generated reverse template was not completely symmetric with the
forward template (which caused an UndeterminableKineticsException in a
recent RMG-Py methylformate job). The fix is to include Cdd as an allowed
atom type in the *3 position in the general templates.

I suspect this is a problem in Intra_R_Add_Exocyclic as well, but I won't
make the change until we encounter the exception explicitly, in case I am
mistaken.