Added several scripts to help with generating viable training sets.
convertPrIMe.py reads one of the local PrIMe databases (e.g.
H_Abstraction/PrIMe.py) and generates files in that family's new
training folder for each unique reaction in the set. It does not copy
the kinetics from PrIMe, however, because with all of the errors in our
PrIMe.py files and the lack of valid temperature range makes the PrIMe
kinetics of limited use.

queryNIST.py is used to query the NIST kinetics database for a given
reaction, as loaded from one of these training files. It constructs an
HTML query, sends the query to kinetics.nist.gov, and parses the HTML
output to extract the kinetics (with temperature range!) and reference
information and saves them back to the training input file.

kineticsTraining.py uses the set of kinetics in one of the training files
to fit the best kinetics as a single modified Arrhenius expression. It
automatically searches for and ignores the really crazy outliers, and can
plot the results for a human to evaluate. This will be especially useful
for those reactions for which many literature references exist. Ultimately
the output should be placed in the training.py file so that we can use it
to train kinetics group additivity values.