Revert "H_Abstraction/training:309 Atoms were mislabeled!"
This commit was the result of a misunderstanding of how labeled atoms
are assigned to products in reaction families that are their own
reverse. In particular, when applying the recipe for such families, we
finish by swapping the necessary labels so that we can apply the recipe
again to go back from products to reactants. (Unfortunately, this
requires some hardcoding.)

This reverts commit c249f31865fe1c6cb71def77678f35499b38060e.