Added script to split PrIMe repositories into NIST and PrIMe counterparts
All entries are first checked for the existence of CAS identifiers for
each species.

Entries with at least one unidentified species are likely to fail on NIST
query, so they are kept in the PrIMe database and will be manually updated

Entries with no unidentified species are consolidated to unique reactions,
then cross-referenced against NIST. Kinetic data (including T and P ranges)
and references are updated appropriately, with the squib entry used as the
entry label.