Restored reference reactions
These PrIMe entries were removed because their NIST counterparts
listed a reference reaction. For most reference reactions, the
rate is clearly not an absolute rate, but rather relative to some
other reaction (i.e., 0.80 * k(rxn2)). This is normally demarked
on the main bibliographic entry page (the squib:0 page), but for
these seven entries, the reference reaction is only listed on the
specific squib page. This, combined with the fact that the values
for these rates appear to be absolutes and match well with other
values for the same reactions, led me to believe these are, in
fact, valid rates.

These restored entries will be thoroughly checked against other
values before migrating to NIST, as well as marked clearly both
in the long description and the commit message that the values
were derived relative to a reference reaction.