Reimported current (4.0-ish) RMG-Java database.
This should be very close to the database released with RMG 4.0. The
PrIMe, NIST, and training files in the kinetics families were not
changed, so any changes made to these since the last database import
are preserved. Any other modifications to the database may need to be
reapplied unless they were also made to the RMG-Java database. The
group additivity values will also need to be retrained, but you don't
necessarily need them to run RMG jobs, view the website, etc.