Redefined Substitution_O to allow for biradical substitution.
The previous definition had an issue with reactions of the form

R-O-R' + R": -> R-O-R". + R'.

which could only be matched in the forward direction (causing an error
message to appear in the log). I've changed the definition to allow for
R-O-R". to also match the appropriate top-level node (now called
O-RR_or_RRrad). This should stop the error message, but I suspect the rate
estimates will be pretty terrible (which could be okay as long as they
aren't sensitive, but who knows).