Updated version of kinetics group additivity values generator.
The kineticsGroups.py script is now again able to generate group
additivity values for the reaction families. Much of the code for this
was moved to KineticsGroups.generateGroupAdditivityValues() in the
RMG-Py codebase. A few other functions -- notably generate() and
createDataSet() -- needed some updating to reflect recent changes in
the KineticsFamily and related classes.