Fixed n value in H_Abstraction NIST entry #1134.
In that reference, the main table gives an n of 5.1, while the discussion
on p. 21 gives a value of 2.5. The latter is the same as the n value for
the analogous reaction to form t-butyl (instead of isobutyl) radical.
The latter is also closer to the preexisting rate rule for this reaction.