Allow *3 to be Cdd in the DoubleBond group as well in Intra_R_Add_Exocyclic
The top level groups all allow Cdd atom type as *3, but we would not
descend the tree to get any decent data, because the double bond group was
defined as Cd only. Now at least we will get to a level two node, although
we have no relevant data below this.