Fix misnamed groups in rules
This only fixed the names of the groups in the kinetics families. It has
not checked whether the groups in groups.py actually match those in
rules.py.

There are still about a dozen or so mistakes that were not easy to figure
out by hand. I found these using a tests I've made for RMG-Py, which I
plan to expand to solve the other cases