Minor fixes to intra_H_migration family
Found using checkWellFormed

-Cs_H_out_Cs2 adjList in rules changed to match group.py. group.py
adjList was more consistent with siblings.

-in R3H_SS_S, some atoms changed from R to R!H. The atoms had two bonds or
a radical, so they had to be by definition

-Cs_H_out_H/OneDe LogicOr was replaced with an AdjList. This was a strange
case where the parent and children were both Groups (Usually I would
expect a given LogicOr group to have LogicOr parent. With the new
flexibility that we have we have programmed into AdjList, I'm not even
sure we need LogicOr anymore). I'm not sure if we should never have this
weird sandwiched LogicOr, but in this case I found I could come up with an
AdjList that fulfills all the correct relationships.