Fix bug with Leave one out test
Previously was calculating Q using natural log, but I want to use log base
10 so that I get orders of magnitude for my comparisons.

Also I was using an averaged temperature of the rule's range for
calculating k. This was when I wanted to calculated the Leave-one-out
value vs the rule's value. Now, I'm pretty sure I want to compare node to
node for future things, so I've just simplified it so that I calculate
everything at 1000K.