Fixes to R_Addition_MultipleBond family
These were found by the checkWellFormed function.

Changes to rule labels rules.py.
For template A-->B (x occurences). 'A' did not exist in groups.py, so the
name was changed to the group 'B' with a matching adjList. 'x' is the
number of rules which used 'A'.

-O_sec_rad --> OJ_sec (1 occurance)
-H_rad does --> HJ (5 occurrences)
-N3t_Ct/H --> N3t_Ct-H (1 occurrence)
-O_pri_rad --> OJ_pri (6 occurrences)
-O_rad/NonDeO --> OJ-Os (13 occurrences)
-Ct/H_N3t --> Ct-H_N3t (2 occurrences)
-N3d/H_N3d/H --> N3d-H_N3d-H (1 occurrence)
-N3d/H_Cds/H2 --> N3d-H_Cds-HH (1 occurrence)
-Cds/H2_N3d --> Cds-HH_N3d (1 occurrence)
-N3t_CtH --> N3t_Ct-H (1 occurrence)

Change to adj lists in rules.py
-Changed radical from 3 to 3Q in "N3t_N3t;CH_quartet" to match group
definition. (CH_quartet only appears in one rule by bbuesser).
-For the top node R_R;YJ, the definitions of R_R and YJ? were both updated
to the more inclusive definition in groups.py (effects top node)
-Ct-Ct_Ct-Cs was found to be wrongly defined in groups.py and rules.py, so
changed all occurrences to the correct definition (19 occurrences)
-Ct-Ct_Ct-H was found to be wrongly defined in groups.py and rules.py, so
changed all occurrences to the correct definition (19 occurrences)

Change to adjlist in groups.py
-CH2_triplet changed to only allow 2T instead of 2S in N3t_N3t;CH2_triplet
(effects 1 rule by bbuesser)
-Ct-Ct_Ct-Cs adj list was definitely wrong. Found because it was a
duplicate of Ct-Cd_Ct-Cs. Changed the Cd in the adj list to a Ct. (appears
in 19 rules)
-Ct-Ct_Ct-H adj list was definitely wrong. Found because it was a
duplicate of Ct-Cd_Ct-H. Changed the Cd in the adj list to a Ct. (appears
in 19 rules)
-Changed Ss atom to more general S atom in SJ. It had the same adj list as
it's only child SsJ. (no rules use Ss)