Fixes to Disproportionation family
-For three rules: changed atom from C to Ct for group Ct_rad/Ct.

-For rule Ct_rad;O_Csrad, the original rule was for C2H + CH2OH --> C2H2 +
CH2O, but the group has been changed to be inclusive for nitrogen. Since
we have an original reaction, I decided to make a training reaction and
remove this rule.

I found the last error serendipitously because the rule's adjList was
mismatched with the groups. I don't have a systematic check for when rules
may not be applicable to expanding rules.