Corrected adjlist for USC-Mech-ii
The changeTripletToSinglet method in rmgpy/molecule/molecule
was used for the singlet radical cases. The method 
changes the multiplicity to 1, but the adjlist still reports
the atom as having 2 unpaired electrons (instead of a lone
pair and an empty shell). I can't remember if this is what
we want of if this should have been corrected.

Should address #50.