Fixes to thermo groups 'group' so that parent-child relationships are correct.
Most fixes were related to parents being more specific than children, so
many C's were converted to CO's or CS's but were accurate groups and
correct parent-child relationships.

Major fix:
-Changing CsJ to include CS atomtypes as well as Cd and CO to now include the
C=S related groups underneath it in the tree.

Additional fixes include:

-Assigning Cs-NNCsH to Cs-NNCsCs which was what it actually described.
 Renaming Cs-N5dN5dCsH to Cs-N5dN5dCsCs and putting it under Cs-NNCsCs
-Creating a new Cs-NNCsH group and putting Cs-(N5dOdOs)(N5dOdOs)CsH under it.
-Changing Cb-CbCbN3s from the old adjlist:

1 * Cb  0 {2,B} {3,B} {4,S}
2   Cb  0 {1,B}
3   Cb  0 {1,B}
4   N3s 0 {1,S}

to the new adjlist so that it would be a proper child of Cb-Cb:

1 * Cb  0 {2,S} {3,D} {4,S}
2   Cb  0 {1,S}
3   Cb  0 {1,D}
4   N3s 0 {1,S}

Hopefully this is what Beat intended.