Added solvation thermo groups for CtOt and OtCt.
The values for these groups are the same as for CdH2 and Od, respectively. This makes it so that the value of the solute descriptors for resonance isomers for carbon monoxide are the same. They are not "real" Platts groups and there is still no Platts group for Ot bonded to anything besides Ct. This way we still catch when there are no Platts groups for something RMG encounters.