Change order of recipe in Silylene_to_Silene family.
Since we were changing the lone pairs first, this led to an incorrect
calculation of the charge on Si. We should probably make it so that
the order we write the recipe does not matter! It happens because updating
lone pairs is the only time we update the charge!
Also, updated group definition so that we can't have 2 lone pairs on Si.