Add convertKineticsLibraryToTrainingReactions script
Specify the kinetics library name below and run the script.
It automatically overwrites the training reactions files it needs to.
 Then you should commit those files.

This script only trains safely. In other words, if a single match
from an RMG family is found, a training reaction is created. Sometimes,
 there are no matches from RMG reaction families, or multiple matches.
This indicates an error that requires manual fixing, and a printout is
 given in the script.